Remove fixtures from database folder and place in top level conftest file, add tests for new session route
